requirements.txt
PyPDF2
openai
python-dotenv
matplotlib
requests
pandas